DRY up the API request methods
Move the top-level API methods into a Base class and genericise the method
names so that they're suitable for different types of authentication.
Http -> Api
