--- lib/cyclid/client/http.rb
@@ -1,48 +1,0 @@
-# Copyright 2016 Liqwyd Ltd.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-require 'uri'
-require 'net/http'
-
-require 'cyclid/client/http/hmac'
-
-module Cyclid
-  # Cyclid client methods
-  module Client
-    # Client HTTP methods
-    module Http
-      # Build a URI for the configured server & required resource
-      def server_uri(path)
-        URI::HTTP.build(host: @config.server,
-                        port: @config.port,
-                        path: path)
-      end
-
-      # Parse, validate & handle response data
-      def parse_response(res)
-        response_data = Oj.load(res.body)
-
-        # Return immediately if the response was an HTTP 200 OK
-        return response_data if res.code == '200'
-
-        @logger.info "server request failed with error ##{res.code}: \
-                     #{response_data['description']}"
-        raise response_data['description']
-      rescue Oj::ParseError => ex
-        @logger.debug "body: #{res.body}\n#{ex}"
-        raise 'failed to decode server response body'
-      end
-    end
-  end
-end
